refactor(server): merge util/ into utils/ and clean comments
