+++ server/app/src/main/kotlin/pos/ambrosia/utils/Exceptions.kt
@@ -1,6 +1,5 @@
 package pos.ambrosia.utils
 
-// Custom exceptions for specific error scenarios
 class InvalidCredentialsException(
     message: String = "Invalid credentials",
 ) : IllegalArgumentException(message)
@@ -13,7 +12,6 @@ class UnauthorizedApiException(
     message: String = "Unauthorized API access",
 ) : SecurityException(message)
 
-// Phoenix Lightning node exceptions
 class PhoenixConnectionException(
     message: String = "Unable to connect to Phoenix Lightning node",
 ) : RuntimeException(message)
